RAG/RAG-gemini/requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-core
langchain-chroma
python-dotenv
streamlit
pypdf
chromadb
